Move Options class to its own file
